requirements.txt
flask
gunicorn
crewai
python-dotenv
pandas